bump.py: exit if the new directory already exists
Test: bump.py 202404 202505 &&  bump.py 202404 202505
Bug: 279809333

+++ tools/bump.py
@@ -44,6 +44,8 @@ class Bump(object):
         self.versions = [e for e in os.listdir(self.current_release_dir) if e.startswith("android-")]
 
     def run(self):
+        if os.path.exists(self.new_release_dir):
+            return
         shutil.copytree(self.current_release_dir, self.new_release_dir)
         for version in self.versions:
             dst = os.path.join(self.new_release_dir, version)
